# ClientSDKs/StoreApi/openapi.g.yaml
openapi: 3.0.1
paths:
...
  /blurb:
    post:
      summary: Add a new blurb to the store
      operationId: createBlurb
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blurb'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blurb'
        405:
          description: Invalid input
      x-lz-gencall: BlurbRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing blurb
      operationId: updateBlurb
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blurb'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blurb'
        400:
          description: Invalid ID supplied
        404:
          description: Blurb not found
        405:
          description: Validation exception
      x-lz-gencall: BlurbRepo.UpdateAsync(callerInfo, body)
  /blurb/listBlurbs:
    get:
      summary: List all blurbs
      operationId: listBlurbs
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blurb'
        400:
          description: Invalid status value
      x-lz-gencall: BlurbRepo.ListAsync(callerInfo)
  /blurb/ListByStatus:
    get:
      summary: List Blurbs by status
      description: Status
      operationId: listBlurbsByStatus
      parameters:
      - name: blurbStatus
        in: query
        required: true
        description: Status value
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blurb'
        400:
          description: Invalid status value
      x-lz-gencall: BlurbRepo.ListBlurbsByStatusAsync(callerInfo, blurbStatus)
  /message:
    post:
      summary: Add a new message to the store
      operationId: createMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        405:
          description: Invalid input
      x-lz-gencall: MessageRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing message
      operationId: updateMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        400:
          description: Invalid ID supplied
        404:
          description: Message not found
        405:
          description: Validation exception
      x-lz-gencall: MessageRepo.UpdateAsync(callerInfo, body)
  /premise:
    post:
      summary: Add a new premise to the store
      operationId: createPremise
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Premise'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Premise'
        405:
          description: Invalid input
      x-lz-gencall: PremiseRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing premise
      operationId: updatePremise
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Premise'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Premise'
        400:
          description: Invalid ID supplied
        404:
          description: Premise not found
        405:
          description: Validation exception
      x-lz-gencall: PremiseRepo.UpdateAsync(callerInfo, body)
  /chat:
    post:
      summary: Add a new chat to the store
      operationId: createChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        405:
          description: Invalid input
      x-lz-gencall: ChatRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing chat
      operationId: updateChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        400:
          description: Invalid ID supplied
        404:
          description: Chat not found
        405:
          description: Validation exception
      x-lz-gencall: ChatRepo.UpdateAsync(callerInfo, body)
...
  '/blurb/{blurbId}':
    get:
      summary: Read blurb by ID
      description: Returns a single blurb
      operationId: readBlurbById
      parameters:
      - name: blurbId
        in: path
        required: true
        description: ID of blurb to return
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blurb'
        400:
          description: Invalid ID supplied
        404:
          description: Blurb not found
      x-lz-gencall: BlurbRepo.ReadAsync(callerInfo, blurbId)
    delete:
      summary: Deletes a blurb
      operationId: deleteBlurb
      parameters:
      - name: blurbId
        in: path
        required: true
        description: Blurb id to delete
        schema:
          type: string
      responses:
        200:
          description: Success
        400:
          description: Invalid ID supplied
        404:
          description: Blurb not found
      x-lz-gencall: BlurbRepo.DeleteAsync(callerInfo, blurbId)
  '/message/{messageId}':
    get:
      summary: Read message by ID
      description: Returns a single message
      operationId: readMessageById
      parameters:
      - name: messageId
        in: path
        required: true
        description: ID of message to return
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        400:
          description: Invalid ID supplied
        404:
          description: Message not found
      x-lz-gencall: MessageRepo.ReadAsync(callerInfo, messageId)
    delete:
      summary: Deletes a message
      operationId: deleteMessage
      parameters:
      - name: messageId
        in: path
        required: true
        description: Message id to delete
        schema:
          type: string
      responses:
        200:
          description: Success
        400:
          description: Invalid ID supplied
        404:
          description: Message not found
      x-lz-gencall: MessageRepo.DeleteAsync(callerInfo, messageId)
  '/message/ListByChatId/{chatId}':
    get:
      summary: List Messages by ChatId
      description: Lists messages by chatId
      operationId: listMessagesByChatId
      parameters:
      - name: chatId
        in: query
        required: true
        description: ChatId message belongs to
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        400:
          description: Invalid status value
      x-lz-gencall: MessageRepo.ListMessagesByChatIdAsync(callerInfo, chatId)
  '/premise/{premiseId}':
    get:
      summary: Read premise by ID
      description: Returns a single premise
      operationId: readPremiseById
      parameters:
      - name: premiseId
        in: path
        required: true
        description: ID of premise to return
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Premise'
        400:
          description: Invalid ID supplied
        404:
          description: Premise not found
      x-lz-gencall: PremiseRepo.ReadAsync(callerInfo, premiseId)
    delete:
      summary: Deletes a premise
      operationId: deletePremise
      parameters:
      - name: premiseId
        in: path
        required: true
        description: Premise id to delete
        schema:
          type: string
      responses:
        200:
          description: Success
        400:
          description: Invalid ID supplied
        404:
          description: Premise not found
      x-lz-gencall: PremiseRepo.DeleteAsync(callerInfo, premiseId)
  '/premise/listByBlurbId/{blurbId}':
    get:
      summary: List Premises by blurbId
      description: List Premiess by blurbId
      operationId: listPremisesByBlurbId
      parameters:
      - name: blurbId
        in: query
        required: true
        description: BlurbId the premise belongs to
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Premise'
        400:
          description: Invalid blurbId value
      x-lz-gencall: PremiseRepo.ListPremisesByBlurbIdAsync(callerInfo, blurbId)
  '/chat/{chatId}':
    get:
      summary: Read chat by ID
      description: Returns a single chat
      operationId: readChatById
      parameters:
      - name: chatId
        in: path
        required: true
        description: ID of chat to return
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        400:
          description: Invalid ID supplied
        404:
          description: Chat not found
      x-lz-gencall: ChatRepo.ReadAsync(callerInfo, chatId)
    delete:
      summary: Deletes a chat
      operationId: deleteChat
      parameters:
      - name: chatId
        in: path
        required: true
        description: Chat id to delete
        schema:
          type: string
      responses:
        200:
          description: Success
        400:
          description: Invalid ID supplied
        404:
          description: Chat not found
      x-lz-gencall: ChatRepo.DeleteAsync(callerInfo, chatId)
  '/chat/listByBlurbId/{blurbId}':
    get:
      summary: List Chats by blurbId
      description: blurbId value chat belongs to
      operationId: listChatsByBlurbId
      parameters:
      - name: blurbId
        in: query
        required: true
        description: BlurbId chat belongs to
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
        400:
          description: Invalid blurbId value
      x-lz-gencall: ChatRepo.ListChatsByBlurbIdAsync(callerInfo, blurbId)
  '/chat/listByPremiseId/{premiseId}':
    get:
      summary: List Chats by premiseId
      description: premiseId value chat belongs to
      operationId: listChatsByPremiseId
      parameters:
      - name: premiseId
        in: query
        required: true
        description: PremiseId chat belongs to
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
        400:
          description: Invalid premiseId value
      x-lz-gencall: ChatRepo.ListChatsByPremiseIdAsync(callerInfo, premiseId)
...
components:
  schemas:
    Blurb:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      x-lz-genrepo: true
    Message:
      type: object
      properties:
        id:
          type: string
        blurbId:
          type: string
        premiseId:
          type: string
        body:
          type: string
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
    Premise:
      type: object
      properties:
        id:
          type: string
        blurbId:
          type: string
        body:
          type: string
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
    Chat:
      type: object
      properties:
        id:
          type: string
        blurbId:
          type: string
        premiseId:
          type: string
        body:
          type: string
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
...